+++ netlify.toml
@@ -8,13 +8,18 @@
 
 [build.environment]
   # Set Node.js version
-  NODE_VERSION = "20.18.0"
+  NODE_VERSION = "20.18.1"
   # Set production environment
   NODE_ENV = "production"
   # Increase timeout for dependency installation
   NETLIFY_TIMEOUT = "600"
   # Force use of Yarn instead of pnpm
   NPM_CONFIG_PACKAGE_MANAGER = "yarn"
+  # Disable Next.js telemetry and SWC download issues
+  NEXT_TELEMETRY_DISABLED = "1"
+  NEXT_PRIVATE_SKIP_SWC_DOWNLOAD = "1"
+  # Skip React version check
+  SKIP_REACT_VERSION_CHECK = "1"
 
 # Build settings
 [build.processing]
