+++ next.config.js
@@ -21,7 +21,6 @@ const nextConfig = {
 	
 	// Bundle optimization
 	experimental: {
-		optimizeCss: true,
 		optimizePackageImports: ['lucide-react', '@radix-ui/react-icons'],
 	},
