+++ package.json
@@ -56,7 +56,7 @@
     "js-yaml": "^4.1.0",
     "jsdom": "^24.1.0",
     "lucide-react": "^0.544.0",
-    "next": "^13.5.6",
+    "next": "^14.2.0",
     "node-cron": "^4.2.1",
     "openai": "^5.11.0",
     "path": "^0.12.7",
@@ -66,6 +66,7 @@
     "react-datepicker": "^8.4.0",
     "react-dom": "^18.3.1",
     "recharts": "^3.1.2",
+    "react-is": "^18.3.1",
     "resend": "^6.0.3",
     "uuid": "^13.0.0",
     "web-vitals": "^5.1.0"
@@ -97,38 +98,7 @@
     "@radix-ui/react-toast": "^1.1.5",
     "@radix-ui/react-tooltip": "^1.0.7",
     "@stripe/stripe-js": "^2.3.0",
-    "@supabase/supabase-js": "^2.39.3",
     "@tanstack/react-query": "^5.17.19",
-    "class-variance-authority": "^0.7.0",
-    "clsx": "^2.1.0",
-    "cmdk": "^0.2.1",
-    "date-fns": "^3.3.1",
-    "embla-carousel-react": "^8.0.0-rc19",
-    "framer-motion": "^11.0.3",
-    "i18next": "^23.7.16",
-    "i18next-browser-languagedetector": "^7.2.0",
-    "input-otp": "^1.1.0",
-    "jspdf": "^2.5.1",
-    "jspdf-autotable": "^3.8.1",
-    "lucide-react": "^0.323.0",
-    "next-themes": "^0.2.1",
-    "react": "^18.2.0",
-    "react-day-picker": "^8.10.1",
-    "react-dom": "^18.2.0",
-    "react-helmet-async": "^2.0.4",
-    "react-hook-form": "^7.50.1",
-    "react-i18next": "^14.1.3",
-    "react-router-dom": "^6.22.0",
-    "recharts": "^2.15.4",
-    "sonner": "^2.0.7",
-    "tailwind-merge": "^2.2.1",
-    "vaul": "^0.9.0",
-    "zod": "^3.22.4"
-  },
-  "devDependencies": {
-    "@commitlint/cli": "^19.8.1",
-    "@commitlint/config-conventional": "^19.8.1",
-    "@eslint/js": "^9.32.0",
     "@testing-library/jest-dom": "^6.6.4",
     "@testing-library/react": "^14.3.1",
     "@types/node": "24.2.1",
@@ -138,16 +108,36 @@
     "@typescript-eslint/eslint-plugin": "^8.39.0",
     "@typescript-eslint/parser": "^8.39.0",
     "autoprefixer": "^10.4.21",
+    "class-variance-authority": "^0.7.0",
+    "clsx": "^2.1.0",
+    "cmdk": "^0.2.1",
+    "date-fns": "^3.3.1",
+    "embla-carousel-react": "^8.0.0-rc19",
     "eslint": "^8.57.0",
     "eslint-config-next": "^15.4.6",
     "eslint-plugin-react-hooks": "^5.2.0",
     "eslint-plugin-react-refresh": "^0.4.20",
     "husky": "^9.1.7",
+    "i18next": "^23.7.16",
+    "i18next-browser-languagedetector": "^7.2.0",
+    "input-otp": "^1.1.0",
+    "jspdf": "^2.5.1",
+    "jspdf-autotable": "^3.8.1",
     "lint-staged": "^16.1.4",
+    "next-themes": "^0.2.1",
     "postcss": "^8.4.32",
+    "react-day-picker": "^8.10.1",
+    "react-helmet-async": "^2.0.4",
+    "react-hook-form": "^7.50.1",
+    "react-i18next": "^14.1.3",
+    "react-router-dom": "^6.22.0",
+    "sonner": "^2.0.7",
+    "tailwind-merge": "^2.2.1",
     "tailwindcss": "^3.4.0",
     "ts-node": "^10.9.2",
-    "typescript": "^5.9.2"
+    "typescript": "^5.9.2",
+    "vaul": "^0.9.0",
+    "zod": "^3.22.4"
   },
   "lint-staged": {
     "**/*.{js,jsx,ts,tsx}": [
