+++ src/utils/seo.tsx
@@ -1,0 +1,113 @@
+import React, { useEffect, useState } from 'react';
+import { getCLS, getFID, getFCP, getLCP, getTTFB } from 'web-vitals';
+
+interface PerformanceMetrics {
+  cls: number;
+  fid: number;
+  fcp: number;
+  lcp: number;
+  ttfb: number;
+}
+
+export const PerformanceMonitor: React.FC = () => {
+  const [metrics, setMetrics] = useState<PerformanceMetrics | null>(null);
+  const [isVisible, setIsVisible] = useState(false);
+
+  useEffect(() => {
+    const metricValues: Partial<PerformanceMetrics> = {};
+
+    getCLS((metric) => {
+      metricValues.cls = metric.value;
+      setMetrics(prev => ({ ...prev, ...metricValues } as PerformanceMetrics));
+    });
+
+    getFID((metric) => {
+      metricValues.fid = metric.value;
+      setMetrics(prev => ({ ...prev, ...metricValues } as PerformanceMetrics));
+    });
+
+    getFCP((metric) => {
+      metricValues.fcp = metric.value;
+      setMetrics(prev => ({ ...prev, ...metricValues } as PerformanceMetrics));
+    });
+
+    getLCP((metric) => {
+      metricValues.lcp = metric.value;
+      setMetrics(prev => ({ ...prev, ...metricValues } as PerformanceMetrics));
+    });
+
+    getTTFB((metric) => {
+      metricValues.ttfb = metric.value;
+      setMetrics(prev => ({ ...prev, ...metricValues } as PerformanceMetrics));
+    });
+  }, []);
+
+  if (!metrics || !isVisible) {
+    return (
+      <button
+        onClick={() => setIsVisible(true)}
+        className="fixed bottom-4 right-4 bg-blue-600 text-white p-2 rounded-full shadow-lg z-50"
+        title="Show Performance Metrics"
+      >
+        📊
+      </button>
+    );
+  }
+
+  const getScoreColor = (value: number, thresholds: { good: number; poor: number }) => {
+    if (value <= thresholds.good) return 'text-green-600';
+    if (value <= thresholds.poor) return 'text-yellow-600';
+    return 'text-red-600';
+  };
+
+  return (
+    <div className="fixed bottom-4 right-4 bg-white dark:bg-gray-800 p-4 rounded-lg shadow-lg z-50 max-w-xs">
+      <div className="flex justify-between items-center mb-2">
+        <h3 className="font-semibold text-sm">Performance Metrics</h3>
+        <button
+          onClick={() => setIsVisible(false)}
+          className="text-gray-500 hover:text-gray-700"
+        >
+          ✕
+        </button>
+      </div>
+      
+      <div className="space-y-2 text-xs">
+        <div className="flex justify-between">
+          <span>LCP:</span>
+          <span className={getScoreColor(metrics.lcp, { good: 2500, poor: 4000 })}>
+            {metrics.lcp.toFixed(0)}ms
+          </span>
+        </div>
+        
+        <div className="flex justify-between">
+          <span>FID:</span>
+          <span className={getScoreColor(metrics.fid, { good: 100, poor: 300 })}>
+            {metrics.fid.toFixed(0)}ms
+          </span>
+        </div>
+        
+        <div className="flex justify-between">
+          <span>CLS:</span>
+          <span className={getScoreColor(metrics.cls, { good: 0.1, poor: 0.25 })}>
+            {metrics.cls.toFixed(3)}
+          </span>
+        </div>
+        
+        <div className="flex justify-between">
+          <span>FCP:</span>
+          <span className={getScoreColor(metrics.fcp, { good: 1800, poor: 3000 })}>
+            {metrics.fcp.toFixed(0)}ms
+          </span>
+        </div>
+        
+        <div className="flex justify-between">
+          <span>TTFB:</span>
+          <span className={getScoreColor(metrics.ttfb, { good: 800, poor: 1800 })}>
+            {metrics.ttfb.toFixed(0)}ms
+          </span>
+        </div>
+      </div>
+    </div>
+  );
+};
